# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CrossBorder AI Demo API",
    "version": "1.2.0"
  },
  "paths": {
...
    "/api/demo/context": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取香港当前时间信息",
        "description": "返回香港当前时间、有效预测范围和轮询间隔。",
        "operationId": "get_demo_context_api_demo_context_get",
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoContextResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/personas": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取本地 Demo 身份",
        "operationId": "get_demo_personas_api_demo_personas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoPersonasResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/v1-model": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取 AI v1 合成数据评估与产物状态",
        "operationId": "get_v1_model_api_demo_v1_model_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ModelResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/v1-readiness": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取 V1 完整 Demo 就绪度",
        "operationId": "get_v1_readiness_api_demo_v1_readiness_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ReadinessResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/audit": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取本地 Demo 写操作审计",
        "operationId": "get_audit_events_api_demo_audit_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventListResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/model-shadow-summary": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取 AI v1 影子观测汇总",
        "operationId": "get_model_shadow_summary_api_demo_model_shadow_summary_get",
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowObservationSummaryResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/v2-readiness": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取 V2 模型训练与晋级就绪度",
        "operationId": "get_v2_readiness_api_demo_v2_readiness_get",
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ReadinessResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/reset": {
      "post": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "重置 Demo 动态数据",
        "description": "删除已保存的反馈、订阅和企业方案，然后重新加载 JSON 种子。",
        "operationId": "reset_demo_api_demo_reset_post",
        "responses": {
          "200": {
            "description": "重置成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoResetResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEvent": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "persona_id": {
            "type": "string",
            "title": "Persona Id"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "request_id",
          "persona_id",
          "organization_id",
          "method",
          "path",
          "status_code",
          "created_at"
        ],
        "title": "AuditEvent"
      },
      "AuditEventListResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "events",
          "total"
        ],
        "title": "AuditEventListResponse"
      },
...
      "DataSourceStatus": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "fallback": {
            "type": "boolean",
            "title": "Fallback"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "data_version": {
            "type": "string",
            "title": "Data Version"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "source",
          "fetched_at",
          "status",
          "fallback",
          "data_version"
        ],
        "title": "DataSourceStatus"
      },
      "DemoContextResponse": {
        "properties": {
          "current_time": {
            "type": "string",
            "format": "date-time",
            "title": "Current Time"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "min_target_time": {
            "type": "string",
            "format": "date-time",
            "title": "Min Target Time"
          },
          "suggested_target_time": {
            "type": "string",
            "format": "date-time",
            "title": "Suggested Target Time"
          },
          "max_target_time": {
            "type": "string",
            "format": "date-time",
            "title": "Max Target Time"
          },
          "poll_interval_seconds": {
            "type": "integer",
            "title": "Poll Interval Seconds"
          }
        },
        "type": "object",
        "required": [
          "current_time",
          "timezone",
          "min_target_time",
          "suggested_target_time",
          "max_target_time",
          "poll_interval_seconds"
        ],
        "title": "DemoContextResponse"
      },
      "DemoPersona": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "organization_name": {
            "type": "string",
            "title": "Organization Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "organization_id",
          "organization_name"
        ],
        "title": "DemoPersona"
      },
      "DemoPersonasResponse": {
        "properties": {
          "default_persona_id": {
            "type": "string",
            "title": "Default Persona Id"
          },
          "personas": {
            "items": {
              "$ref": "#/components/schemas/DemoPersona"
            },
            "type": "array",
            "title": "Personas"
          }
        },
        "type": "object",
        "required": [
          "default_persona_id",
          "personas"
        ],
        "title": "DemoPersonasResponse"
      },
      "DemoResetResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "seeded": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Seeded"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "seeded",
          "message"
        ],
        "title": "DemoResetResponse"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "ShadowObservationPortSummary": {
        "properties": {
          "port_id": {
            "type": "string",
            "title": "Port Id"
          },
          "port_name": {
            "type": "string",
            "title": "Port Name"
          },
          "observation_count": {
            "type": "integer",
            "title": "Observation Count"
          },
          "average_difference_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Difference Minutes"
          },
          "average_absolute_difference_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Absolute Difference Minutes"
          }
        },
        "type": "object",
        "required": [
          "port_id",
          "port_name",
          "observation_count"
        ],
        "title": "ShadowObservationPortSummary"
      },
      "ShadowObservationSummaryResponse": {
        "properties": {
          "total_observations": {
            "type": "integer",
            "title": "Total Observations"
          },
          "available_observations": {
            "type": "integer",
            "title": "Available Observations"
          },
          "unavailable_observations": {
            "type": "integer",
            "title": "Unavailable Observations"
          },
          "latest_observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Observed At"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/ShadowObservationPortSummary"
            },
            "type": "array",
            "title": "Ports"
          }
        },
        "type": "object",
        "required": [
          "total_observations",
          "available_observations",
          "unavailable_observations",
          "ports"
        ],
        "title": "ShadowObservationSummaryResponse"
      },
...
      "V1ModelResponse": {
        "properties": {
          "artifact_available": {
            "type": "boolean",
            "title": "Artifact Available"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "synthetic_only": {
            "type": "boolean",
            "title": "Synthetic Only"
          },
          "dataset": {
            "additionalProperties": true,
            "type": "object",
            "title": "Dataset"
          },
          "split": {
            "additionalProperties": true,
            "type": "object",
            "title": "Split"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "promotion": {
            "additionalProperties": true,
            "type": "object",
            "title": "Promotion"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "artifact_available",
          "model_version",
          "synthetic_only",
          "dataset",
          "split",
          "metrics",
          "features",
          "promotion",
          "limitations"
        ],
        "title": "V1ModelResponse"
      },
      "V1ReadinessCheck": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "name",
          "passed",
          "detail"
        ],
        "title": "V1ReadinessCheck"
      },
      "V1ReadinessResponse": {
        "properties": {
          "demo_ready": {
            "type": "boolean",
            "title": "Demo Ready"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/V1ReadinessCheck"
            },
            "type": "array",
            "title": "Checks"
          },
          "adapter_modes": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Adapter Modes"
          }
        },
        "type": "object",
        "required": [
          "demo_ready",
          "checks",
          "adapter_modes"
        ],
        "title": "V1ReadinessResponse"
      },
      "V2ReadinessCheck": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "actual": {
            "type": "integer",
            "title": "Actual"
          },
          "required": {
            "type": "integer",
            "title": "Required"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          }
        },
        "type": "object",
        "required": [
          "name",
          "actual",
          "required",
          "passed"
        ],
        "title": "V2ReadinessCheck"
      },
      "V2ReadinessLabelSource": {
        "properties": {
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "label_count": {
            "type": "integer",
            "title": "Label Count"
          }
        },
        "type": "object",
        "required": [
          "source_type",
          "label_count"
        ],
        "title": "V2ReadinessLabelSource"
      },
      "V2ReadinessPort": {
        "properties": {
          "port_id": {
            "type": "string",
            "title": "Port Id"
          },
          "label_count": {
            "type": "integer",
            "title": "Label Count"
          }
        },
        "type": "object",
        "required": [
          "port_id",
          "label_count"
        ],
        "title": "V2ReadinessPort"
      },
      "V2ReadinessResponse": {
        "properties": {
          "experiment_ready": {
            "type": "boolean",
            "title": "Experiment Ready"
          },
          "production_promotion_ready": {
            "type": "boolean",
            "title": "Production Promotion Ready"
          },
          "label_count": {
            "type": "integer",
            "title": "Label Count"
          },
          "linked_feedback_count": {
            "type": "integer",
            "title": "Linked Feedback Count"
          },
          "excluded_feedback_count": {
            "type": "integer",
            "title": "Excluded Feedback Count"
          },
          "label_sources": {
            "items": {
              "$ref": "#/components/schemas/V2ReadinessLabelSource"
            },
            "type": "array",
            "title": "Label Sources"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/V2ReadinessPort"
            },
            "type": "array",
            "title": "Ports"
          },
          "distinct_dates": {
            "type": "integer",
            "title": "Distinct Dates"
          },
          "hour_slices": {
            "type": "integer",
            "title": "Hour Slices"
          },
          "data_versions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Versions"
          },
          "statistical_mae_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statistical Mae Minutes"
          },
          "shadow_mae_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shadow Mae Minutes"
          },
          "shadow_labeled_count": {
            "type": "integer",
            "title": "Shadow Labeled Count"
          },
          "time_split": {
            "additionalProperties": true,
            "type": "object",
            "title": "Time Split"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/V2ReadinessCheck"
            },
            "type": "array",
            "title": "Checks"
          },
          "data_sources": {
            "items": {
              "$ref": "#/components/schemas/DataSourceStatus"
            },
            "type": "array",
            "title": "Data Sources"
          },
          "coverage_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Coverage Warnings"
          },
          "production_blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Production Blockers"
          }
        },
        "type": "object",
        "required": [
          "experiment_ready",
          "production_promotion_ready",
          "label_count",
          "linked_feedback_count",
          "excluded_feedback_count",
          "label_sources",
          "ports",
          "distinct_dates",
          "hour_slices",
          "data_versions",
          "shadow_labeled_count",
          "time_split",
          "checks",
          "data_sources",
          "coverage_warnings",
          "production_blockers"
        ],
        "title": "V2ReadinessResponse"
      },
...
    }
  }
}